+++ lib/chainsync/chainsync/db/base/__init__.py
@@ -1,5 +1,6 @@
 """Generic database utilities"""
 
+from .convert_scaled_value_to_decimal import convert_scaled_value_to_decimal
 from .interface import (
     PostgresConfig,
     TableWithBlockNumber,
@@ -14,4 +15,3 @@ from .interface import (
     query_tables,
 )
 from .schema import Base, UserMap
-from .utils import convert_scaled_value_to_decimal
