+++ lib/chainsync/chainsync/db/base/convert_scaled_value_to_decimal.py
